Move GCMRecord to separate file.
